+++ src/autodiff.jl
@@ -31,109 +31,6 @@ any nonlinear constraint ``h(x)``.
 """
 abstract type AbstractHessian end
 
-abstract type AbstractAdjointStack{VT} end
-
-"""
-    jacobian!(form::AbstractFormulation, jac::AutoDiff.AbstractJacobian, x)
-
-Update inplace the Jacobian ``J`` stored inside `jac` at a new point `x`.
-"""
-function jacobian! end
-
-"""
-    adj_hessian_prod!(form::AbstractFormulation, H::AutoDiff.AbstractHessian, hv, x, λ, v)
-
-Compute the adjoint-Hessian-vector product ``λ^⊤ H v`` at a given
-point `x`, and store the result inplace in vector `hv`.
-
-"""
-function adj_hessian_prod! end
-
-"""
-    AutoDiff.Jacobian <: AbstractJacobian
-
-Creates an object to compute the Jacobian with ForwardDiff.
-
-### Attributes
-
-* `func::Func`: base function to differentiate
-* `var::Union{State,Control}`: specify whether we are differentiating w.r.t. the state or the control.
-* `J::SMT`: Sparse uncompressed Jacobian to be used by linear solver. This is either of type `SparseMatrixCSC` or `CuSparseMatrixCSR`.
-* `compressedJ::MT`: Dense compressed Jacobian used for updating values through AD either of type `Matrix` or `CuMatrix`.
-* `coloring::VI`: Row coloring of the Jacobian.
-* `t1sseeds::VP`: The seeding vector for AD built based on the coloring.
-* `t1sF::VD`: Output array of active (AD) type.
-* `x::VT`: Input array of passive type. This includes both state and control.
-* `t1sx::VD`: Input array of active type.
-* `map::VI`: State and control mapping to array `x`
-* `varx::SubT`: View of `map` on `x`
-* `t1svarx::SubD`: Active (AD) view of `map` on `x`
-
-"""
-struct Jacobian{Func, VI, VT, MT, SMT, VP, VD, SubT, SubD, JacT} <: AbstractJacobian
-    func::Func
-    J::SMT
-    compressedJ::MT
-    coloring::VI
-    t1sseeds::VP
-    t1sF::VD
-    x::VT
-    t1sx::VD
-    map::VI
-    # Cache views on x and its dual vector to avoid reallocating on the GPU
-    varx::SubT
-    t1svarx::SubD
-end
-
-function Base.show(io::IO, jacobian::Jacobian)
-    println(io, "A AutoDiff Jacobian for $(jacobian.func)")
-    ncolor = size(jacobian.compressedJ, 1)
-    print(io, "Number of Jacobian colors: ", ncolor)
-end
-
-"""
-    AutoDiff.ConstantJacobian <: AbstractJacobian
-
-Creates a constant Jacobian object for a linear function ``h(x)``.
-Using a `ConstantJacobian` object allows to avoid computing
-the full Jacobian with AutoDiff when it is not necessary.
-"""
-struct ConstantJacobian{SMT} <: AbstractJacobian
-    J::SMT
-end
-
-"""
-    AutoDiff.Hessian
-
-Creates an object for computing Hessian adjoint tangent projections.
-
-* `func::Func`: base function to differentiate.
-* `host_t1sseeds::VHP`: Seeding vector for seeding on the host.
-* `t1sseeds::VP`: The seeding vector for AD built based on the coloring.
-* `x::VT`: Input array of passive type. This includes both state and control.
-* `t1sF::VD`: Output array of active (AD) type.
-* `∂t1sF::VD`: Adjoint of the output array.
-* `t1sx::VD`: Input array of active type.
-* `∂t1sx::VD`: Adjoint of the input array.
-* `map::VI`: State and control mapping to array `x`
-* `varx::SubT`: View of `map` on `x`
-* `t1svarx::SubD`: Active (AD) view of `map` on `x`
-* `buffer::Buff`: cache for computing the adjoint (could be `Nothing`)
-"""
-struct Hessian{Func, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, Buff} <: AbstractHessian
-    func::Func
-    host_t1sseeds::T1 # Needed because seeds have to be created on the host
-    t1sseeds::T2
-    x::T3
-    t1sF::T4
-    ∂t1sF::T5
-    t1sx::T6
-    ∂t1sx::T7
-    map::T8
-    varx::T9
-    t1svarx::T10
-    buffer::Buff
-end
 
 # Cache for adjoint
 """
